use textSectionActions
